Increased performance of the database significantly.
I also installed the Django debug toolbar so that we can inspect SQL
calls (and their execution times) if need be.

select_related is able to “follow” foreign key relationships,
drastically cutting down on the number of SQL queries our code
performed. If you take a look now (by opening
http://127.0.0.1:8000/sessions/courses/?id=1&id=3&id=37&id=657), you
can see that there is a call being made for every attendee. I’m
thinking that prefetch_related() may be able to fix this, but I can’t
for the life of me figure out how.