Started work on off-campus locations
This is a hacky server-side change to format off-campus location data
into a format that the client can process relatively easy. Essentially,
the only thing that had to be done was to take all offCampusSession
data and wrap it within a location {} dictionary so that the returned
JSON format matches that of onCampusSessions with only a few minor
changes. These minor differences (such as missing room numbers) are
then handled individually on the client.